// packages/2-mongo-family/1-foundation/mongo-contract/test/fixtures/orm-contract.d.ts
export type CodecTypes = MongoCodecTypes;

export type HomeAddressOutput = {
  readonly city: CodecTypes['mongo/string@1']['output'];
  readonly country: CodecTypes['mongo/string@1']['output'];
};
export type HomeAddressInput = {
  readonly city: CodecTypes['mongo/string@1']['input'];
  readonly country: CodecTypes['mongo/string@1']['input'];
};
export type FieldOutputTypes = {
  readonly __unbound__: {
    readonly Address: {
      readonly street: CodecTypes['mongo/string@1']['output'];
      readonly city: CodecTypes['mongo/string@1']['output'];
      readonly zip: CodecTypes['mongo/string@1']['output'];
    };
    readonly Bug: { readonly severity: CodecTypes['mongo/string@1']['output'] };
    readonly Comment: {
      readonly _id: CodecTypes['mongo/objectId@1']['output'];
      readonly text: CodecTypes['mongo/string@1']['output'];
      readonly createdAt: CodecTypes['mongo/date@1']['output'];
    };
    readonly Feature: {
      readonly priority: CodecTypes['mongo/string@1']['output'];
      readonly targetRelease: CodecTypes['mongo/string@1']['output'];
    };
    readonly Task: {
      readonly _id: CodecTypes['mongo/objectId@1']['output'];
      readonly title: CodecTypes['mongo/string@1']['output'];
      readonly type: CodecTypes['mongo/string@1']['output'];
      readonly assigneeId: CodecTypes['mongo/objectId@1']['output'];
    };
    readonly User: {
      readonly _id: CodecTypes['mongo/objectId@1']['output'];
      readonly name: CodecTypes['mongo/string@1']['output'];
      readonly email: CodecTypes['mongo/string@1']['output'];
      readonly loginCount: CodecTypes['mongo/int32@1']['output'];
      readonly tags: ReadonlyArray<CodecTypes['mongo/string@1']['output']>;
      readonly homeAddress: HomeAddressOutput | null;
    };
  };
};
export type FieldInputTypes = {
  readonly __unbound__: {
    readonly Address: {
      readonly street: CodecTypes['mongo/string@1']['input'];
      readonly city: CodecTypes['mongo/string@1']['input'];
      readonly zip: CodecTypes['mongo/string@1']['input'];
    };
    readonly Bug: { readonly severity: CodecTypes['mongo/string@1']['input'] };
    readonly Comment: {
      readonly _id: CodecTypes['mongo/objectId@1']['input'];
      readonly text: CodecTypes['mongo/string@1']['input'];
      readonly createdAt: CodecTypes['mongo/date@1']['input'];
    };
    readonly Feature: {
      readonly priority: CodecTypes['mongo/string@1']['input'];
      readonly targetRelease: CodecTypes['mongo/string@1']['input'];
    };
    readonly Task: {
      readonly _id: CodecTypes['mongo/objectId@1']['input'];
      readonly title: CodecTypes['mongo/string@1']['input'];
      readonly type: CodecTypes['mongo/string@1']['input'];
      readonly assigneeId: CodecTypes['mongo/objectId@1']['input'];
    };
    readonly User: {
      readonly _id: CodecTypes['mongo/objectId@1']['input'];
      readonly name: CodecTypes['mongo/string@1']['input'];
      readonly email: CodecTypes['mongo/string@1']['input'];
      readonly loginCount: CodecTypes['mongo/int32@1']['input'];
      readonly tags: ReadonlyArray<CodecTypes['mongo/string@1']['input']>;
      readonly homeAddress: HomeAddressInput | null;
    };
  };
};
// ...
export type TypeMaps = MongoTypeMaps<CodecTypes, FieldOutputTypes, FieldInputTypes>;
// ...
export type Contract = MongoContractWithTypeMaps<ContractBase, TypeMaps>;